refactor(core): split model unit tests by concern
Move the 1.8k-line model/tests.rs dump into
tests/{progress,download,variant,ane}.rs so coverage sits next to the
modules it exercises.
